+++ src/main/java/ru/vyarus/dropwizard/guice/bundle/lookup/PropertyBundleLookup.java
@@ -1,10 +1,8 @@
 package ru.vyarus.dropwizard.guice.bundle.lookup;
 
-import com.google.common.base.Joiner;
-import com.google.common.base.Splitter;
-import com.google.common.collect.Lists;
 import ru.vyarus.dropwizard.guice.bundle.GuiceyBundleLookup;
 import ru.vyarus.dropwizard.guice.module.installer.bundle.GuiceyBundle;
+import ru.vyarus.dropwizard.guice.module.installer.util.PropertyUtils;
 
 import java.util.Collections;
 import java.util.List;
@@ -30,19 +28,7 @@ public class PropertyBundleLookup implements GuiceyBundleLookup {
 
     @Override
     public List<GuiceyBundle> lookup() {
-        final String prop = System.getProperty(BUNDLES_PROPERTY);
-        List<GuiceyBundle> res = Collections.emptyList();
-        if (prop != null) {
-            final Iterable<String> classes = Splitter.on(',').omitEmptyStrings().trimResults().split(prop);
-            try {
-                res = toInstances(toClasses(classes));
-            } catch (Exception e) {
-                throw new IllegalStateException(String.format(
-                        "Failed to parse bundles system property '%s' value: '%s'",
-                        BUNDLES_PROPERTY, prop), e);
-            }
-        }
-        return res;
+        return PropertyUtils.getProperty(BUNDLES_PROPERTY, Collections.emptyMap());
     }
 
     /**
@@ -53,32 +39,6 @@ public class PropertyBundleLookup implements GuiceyBundleLookup {
      */
     @SafeVarargs
     public static void enableBundles(final Class<? extends GuiceyBundle>... bundles) {
-        final String prop = Joiner.on(',').join(toStrings(Lists.newArrayList(bundles)));
-        System.setProperty(BUNDLES_PROPERTY, prop);
-    }
-
-    private static List<String> toStrings(final Iterable<Class<? extends GuiceyBundle>> list) {
-        final List<String> res = Lists.newArrayList();
-        for (Class cls : list) {
-            res.add(cls.getName());
-        }
-        return res;
-    }
-
-    @SuppressWarnings("unchecked")
-    private List<Class<GuiceyBundle>> toClasses(final Iterable<String> list) throws Exception {
-        final List<Class<GuiceyBundle>> res = Lists.newArrayList();
-        for (String cls : list) {
-            res.add((Class<GuiceyBundle>) Class.forName(cls));
-        }
-        return res;
-    }
-
-    private List<GuiceyBundle> toInstances(final Iterable<Class<GuiceyBundle>> list) throws Exception {
-        final List<GuiceyBundle> res = Lists.newArrayList();
-        for (Class<GuiceyBundle> cls : list) {
-            res.add(cls.newInstance());
-        }
-        return res;
+        PropertyUtils.setProperty(BUNDLES_PROPERTY, bundles);
     }
 }