+++ src/main/java/ru/vyarus/dropwizard/guice/hook/GuiceyConfigurationHook.java
@@ -16,6 +16,16 @@ import ru.vyarus.dropwizard.guice.test.spock.UseGuiceyHooks;
  * <p>
  * Hooks are thread-scoped: it is assumed that registration thread is the same thread where application
  * will start.
+ * <p>
+ * Hooks could be enabled with a system property "guicey.hooks". Property value must contain comma-separated list of
+ * complete hook class names. Each hook in list must have default no-args constructor. Example:
+ * {@code -Dguicey.hooks=com.foo.MyHook1,com.foo.MyHook2}. Aliases may be assigned to simplify hooks enabling
+ * {@link GuiceBundle.Builder#hookAlias(String, Class)}.
+ * <p>
+ * Enabling hooks from system property may be used for enabling reporting or additional tooling on already
+ * compiled applications. For example, bundled {@link DiagnosticHook} could enable guicey diagnostic reports (enabled
+ * during development with print* methods on {@link GuiceBundle}) with system property:
+ * {@code -Dguicey.hooks=diagnostic}.
  *
  * @author Vyacheslav Rusakov
  * @see GuiceyHooksRule
@@ -28,22 +38,22 @@ public interface GuiceyConfigurationHook {
     /**
      * Configuration is applied just after manual configuration (through bundle's builder in application class).
      * <p>
-     * {@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder} contains special methods for test support:
+     * {@link GuiceBundle.Builder} contains special methods for test support:
      * <ul>
      * <li>Generic disable:
-     * {@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder#disable(java.util.function.Predicate[])}
+     * {@link GuiceBundle.Builder#disable(java.util.function.Predicate[])}
      * </li>
      * <li>Direct disable* method, for example
-     * {@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder#disableExtensions(Class[])}</li>
+     * {@link GuiceBundle.Builder#disableExtensions(Class[])}</li>
      * <li>Guice bindings override:
-     * {@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder#modulesOverride(com.google.inject.Module...)}</li>
+     * {@link GuiceBundle.Builder#modulesOverride(com.google.inject.Module...)}</li>
      * </ul>
      * All other configuration options are also available, so it is possible to register extra extensions, bundles etc
-     * or modify guicey options ({@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder#option(Enum, Object)}).
+     * or modify guicey options ({@link GuiceBundle.Builder#option(Enum, Object)}).
      * <p>
      * All configuration items, registered with hook will be scoped as {@link GuiceyConfigurationHook}
      * instead of {@link io.dropwizard.Application} and so will be clearly distinguishable in configuration logs
-     * ({@link ru.vyarus.dropwizard.guice.GuiceBundle.Builder#printDiagnosticInfo()}).
+     * ({@link GuiceBundle.Builder#printDiagnosticInfo()}).
      *
      * @param builder just created bundle's builder
      * @see GuiceyHooksRule for more information